chore (refactoring): rename runtime module
