+++ server/pkg/index.go
@@ -1,0 +1,5 @@
+package core
+
+import (
+	_ "github.com/mickael-kerjean/filestash/server/pkg/sqlite"
+)
