+++ server/pkg/workflow/trigger.go
@@ -1,0 +1,5 @@
+package workflow
+
+import (
+	_ "github.com/mickael-kerjean/filestash/server/pkg/workflow/trigger"
+)
