--- server/runtime/index.go
@@ -1,5 +1,0 @@
-package runtime
-
-import (
-	_ "github.com/mickael-kerjean/filestash/server/runtime/sqlite"
-)
