--- server/workflow/trigger.go
@@ -1,5 +1,0 @@
-package workflow
-
-import (
-	_ "github.com/mickael-kerjean/filestash/server/workflow/trigger"
-)
